pyproject.toml
[project]
name = "match-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "datapizza-ai>=0.0.9",
    "django-environ>=0.12.1",
    "django-redis>=6.0.0",
    "pgvector>=0.4.2",
    "redis>=7.1.1",
]
